Creted the SholdDeleteTool test

+++ src/main/java/com/toolshopmanager/domain/services/tools/DeleteTool.java
@@ -1,11 +1,13 @@
 package com.toolshopmanager.domain.services.tools;
 
+import com.toolshopmanager.domain.entities.tool.Tool;
 import com.toolshopmanager.domain.services.tools.dtos.DeleteToolDTO;
 
+import java.util.Optional;
 import java.util.UUID;
 
 public class DeleteTool {
-    private ToolRepository toolRepository;
+    private final ToolRepository toolRepository;
 
     public DeleteTool(ToolRepository toolRepository) {
         this.toolRepository = toolRepository;
@@ -13,6 +15,12 @@ public class DeleteTool {
 
     public void perform(DeleteToolDTO toolData) {
         UUID toolId = UUID.fromString(toolData.id());
+        Optional<Tool> tool = this.toolRepository.findById(toolId);
+
+        if (tool.isEmpty()) {
+            throw new IllegalArgumentException("Tool is not exist");
+        }
+
         this.toolRepository.delete(toolId);
     }
 }
